Added HomeController and UsersController.
HomeController has following methods:
- index method which maps HTTP GET requests to the root URL and returns the ModelAndView object to the dispatcher servlet, which then resolves and renders the view.
- home method which maps HTTP GET requests to the "/home" URL, accepts an optional request parameter "show-ad" and returns the view name "home".

UsersController has following methods:
- login method which maps HTTP GET requests to the "/login" URL and returns the view name "user/login".
- doLogin method which handles POST requests to the /login URL. The method accepts two parameters @Valid UserLoginDto loginDto(binds the request body to a UserLoginDto
object and applies validation constraints defined in the UserLoginDto class) and BindingResult bindingResult(captures the results of the validation and binding. It holds validation errors, if any).
The method checks if the username is "admin" and password is "secure" and if both conditions are met, it considers the login successful.
If the login is successful, it creates a ModelAndView object and sets the view name to "redirect:/home", which will redirect the user to the "/home" URL. If the login credentials are incorrect,
it collects validation errors from the BindingResult object.  The errors are mapped to a list of strings each containing the object name and the default message. If both conditions above
are not met, a new ModelAndView object is created with the view name set to "user/login". The errors list is added to the model under the key "errors".
Finally, the method returns the ModelAndView object, which will render the login page with the error messages.
- register method which maps HTTP GET requests to the "/register" URL and returns the view name "user/register".
- doRegister method which maps HTTP POST requests to the "/register" URL. Accepts @Valid UserLoginDto loginDto which binds the request body to a UserLoginDto object
and applies validation constraints defined in the UserLoginDto class. The UserService register method is called  which returns boolean representing whether the registration is successful.
If so, returns the view name "redirect: /users/login" which which will redirect the user to the "/users/login" URL, else returns the view name "user/register".
- getUserInfo method which maps HTTP GET requests to the URL pattern "/info/{id}". {id} is a path variable, meaning the value in this part of the URL will be captured
and passed to the method. @PathVariable("id") long id: This annotation binds the path variable {id} from the URL to the method parameter id.
The id is then available as a long type within the method. eturns the view name "home".