Changed HomeController and UsersController
Added new method named 'logout' in HomeController.

In UsersController 'doLogin' and 'doRegister' methods are changed.
- In 'doLogin', the if statement is updated, and the ModelAndView object is now accepted externally.
- The 'doRegister' method now accepts two additional parameters: the ModelAndView object, which is now accepted externally,
and a BindingResult object. The updated if statement includes validation of register credentials. Any validation errors collected
from the BindingResult object are mapped to a list of strings, each containing the object name and its default message. If
registration is successful and credentials are valid, the ModelAndView object is set with the view name 'redirect:/users/login',
rendering the 'login' page. If conditions are not met, the errors list is added to the model under the key 'errors'. The ModelAndView
object is then set with the view name 'user/register', and the method returns the ModelAndView object, rendering the register page
with the error messages.