Added ImportController
Has the following methods:
- 'mainView' - handles the GET request and returns the view name "xml/import-xml".
Checks if companies, projects, and employees have been imported, stores results
in an array and adds the array to the model under the key 'areImported', making it accessible in the view.
'areImported' attribute conditionally displays content based on the import statuses.
- 'getImportCompanies' - handles the GET request and returns the view name "xml/import-companies".
The details of companies obtained from 'companyService.getXmlInfo()' are addded to the model with the key 'companies'.
- 'postCompanies' - handles POST request, calls the 'importCompanies' method of the 'companyService' to perform
the import operation and redirects the user to the "/import/xml" page after the import operation is completed.
- 'getImportProjects' - handles the GET request and returns the view name "xml/import-projects".
The details of projects obtained from 'projectService.getXmlInfo()' are addded to the model with the key 'projects'.
- 'postProjects' - handles POST request, calls the 'importProjects' method of the 'projectService' to perform
the import operation and redirects the user to the "/import/xml" page after the import operation is completed.
- 'getImportEmployees' - handles the GET request and returns the view name "xml/import-employees".
The details of projects obtained from 'employeeService.getXmlInfo()' are addded to the model with the key 'employees'.
- 'postEmployees' - handles POST request, calls the 'importEmployees' method of the 'employeeService' to perform
the import operation and redirects the user to the "/import/xml" page after the import operation is completed.