Added ExportController
Has the following methods:
- 'getAllFinishedProjects' - handles the GET request and returns the view name "export/export-project-if-finished".
Calls 'getExportStringForAllFinishedProjects' method of the 'projectService' which returns a string representation
of all finished projects and adds the string to the model under the key 'projectsIfFinished'.
- 'getAllEmployeesAbove25' - handles the GET request and returns the view name "export/export-employees-with-age".
Calls 'getExportStringForAllEmployeesAbove25Years' method of the 'employeeService' which returns a string representation
of all employees above 25 years old and adds the string to the model under the key 'employeesAbove'.